requirements-gpu.txt
# GPU-Optimized Requirements for Pipecat Voice Pipeline

# Core Framework
pipecat-ai>=0.0.30
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0

# STT (Speech-to-Text)
faster-whisper>=0.10.0
openai-whisper  # Optional, for comparison
ctranslate2>=3.24.0  # GPU acceleration for Whisper

# LLM
groq>=0.4.0

# TTS (Text-to-Speech)
kokoro>=0.3.4
edge-tts>=6.1.0
pydub>=0.25.0  # For audio conversion
azure-cognitiveservices-speech  # Optional

# Audio Processing
numpy>=1.24.0
scipy>=1.11.0
soundfile>=0.12.0

# GPU Support
nvidia-ml-py3>=7.352.0  # NVIDIA GPU monitoring

# Configuration
pyyaml>=6.0.1

# Optional: Local GPU LLM
# vllm  # Uncomment for local LLM support
# transformers>=4.35.0
# accelerate

# Optional: Advanced VAD
# silero-vad  # Uncomment for Silero VAD

# Monitoring (optional)
# prometheus-client
# psutil
